更新版本和更新代码

+++ Managed/main/IGroupProber.cs
@@ -2,9 +2,9 @@ using System.Collections.Generic;
 
 public interface IGroupProber
 {
-	void Occupy(object prober, int serial_no, IEnumerable<int> cells);
+	void Occupy(object prober, short serial_no, IEnumerable<int> cells);
 
-	void SetValidSerialNos(object prober, int previous_serial_no, int serial_no);
+	void SetValidSerialNos(object prober, short previous_serial_no, short serial_no);
 
 	bool ReleaseProber(object prober);
 }
